requirements.txt
telethon
python-dotenv
supabase
requests
